disable tests pending doc review
